src/requirements.txt
PyQt5
python-dateutil

## for ics files
icalendar
pidfile
pqueue
filelock
watchdog

## for iCalendar client/server functionality
radicale
caldav
